pyproject.toml
[project]
name = "poc-rag-gcp-public"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "mkdocs-material>=9.7.6",
    "pymdown-extensions>=10.21",
    "streamlit>=1.45.0",
    "google-cloud-firestore>=2.25.0",
    "vertexai>=1.0.0",
    "google-cloud-discoveryengine>=0.13.0",
    "langchain-text-splitters>=0.3.0",
    "python-dotenv>=1.1.0",
]
